+++ packages/tsurlfilter/test/filterlist/scanner/rule-storage-scanner.test.ts
@@ -81,3 +81,25 @@ describe('RuleStorageScanner Test', () => {
         expect(storageScanner.scan()).toBeFalsy();
     });
 });
+
+// Note: this is needed because some special filter lists have negative IDs,
+// like the internal Stealth Mode's filter list which has ID -1.
+describe('Check negative filter list ID', () => {
+    it('works with negative filter id', () => {
+        const filterList = '||example.org';
+
+        const lineReader = new StringLineReader(filterList);
+        const scanner = new RuleScanner(lineReader, -1, {
+            scannerType: ScannerType.All,
+            ignoreCosmetic: false,
+        });
+
+        const storageScanner = new RuleStorageScanner([scanner]);
+
+        expect(storageScanner.scan()).toBeTruthy();
+        expect(storageScanner.getRule()).toBeTruthy();
+        expect(storageScanner.getRule()!.rule.getFilterListId()).toBe(-1);
+
+        expect(storageScanner.scan()).toBeFalsy();
+    });
+});