+++ packages/tswebextension/CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version changes -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## [1.0.30] - 2024-07-17
+
+### Changed
+
+- Updated `@adguard/tsurlfilter` to `v2.2.22`.
+
+[1.0.30]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v1.0.30
+
 ## [1.0.29] - 2024-07-12
 
 ### Changed